Use less custom channel types

+++ redhorizon-filetypes/source/nz/net/ultraq/redhorizon/filetypes/ArchiveFileEntry.groovy
@@ -36,5 +36,5 @@ interface ArchiveFileEntry {
 	 * 
 	 * @return Size of the entry.
 	 */
-	long getSize()
+	int getSize()
 }
